chess_bot/requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pywin32>=306
pynput>=1.7.6
